% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nei.efdr}
\alias{nei.efdr}
\title{Find wavelet neighbourhood}
\usage{
nei.efdr(Z, wf = "la8", J = 3, b = 11, parallel = FALSE)
}
\arguments{
\item{b}{number of neighbours per wavelet}

\item{parallel}{flag to indicate whether multi-threading should be used or not}

\item{dwt}{object of class \code{dwt.2d}}
}
\value{
matrix of size \code{N} by \code{b}
}
\description{
Given a wavelet basis from object \code{waveslim}, this function returns a matrix of size \code{N} by \code{b}
where \code{N} is the number of wavelets and \code{b} is the number of neighbours per wavelet. Two wavelets are deemed
to be neighbours according to the metric of Shen, Huang and Cressie (2002). The distance metric is a function of  the
spatial separation, the resolution and the orientation.
}
\examples{
image <- matrix(rnorm(64),8,8)
}
\references{
Shen, Xiaotong, Hsin-Cheng Huang, and Noel Cressie. "Nonparametric hypothesis testing for a spatial signal." Journal of the American Statistical Association 97.460 (2002): 1122-1140.
}
\keyword{neighbourhood}
\keyword{wavelets,}

