% Generated by roxygen2 (4.0.2): do not edit by hand
\name{regrid}
\alias{regrid}
\title{Regrid ir/regular data}
\usage{
regrid(df, n1 = 128, n2 = n1, idp = 0.5, nmax = 7)
}
\arguments{
\item{df}{data frame with fields \code{x}, \code{y} and \code{z}}

\item{n1}{image length in pixels}

\item{n2}{image height in pixels}

\item{idp}{the inverse distance power}

\item{nmax}{the number of nearest neighbours to consider when interpolating}
}
\value{
data frame with \code{x,y} as gridded values
}
\description{
Given a data frame with fields \code{x, y} and \code{z}, \code{regrid} returns a data frame with
fields \code{x, y} and \code{z}, this time with \code{x, y} arranged on a regular grid of size \code{n} by
\code{n}.
}
\details{
The function overlays a grid over the data. The cells are constructed evenly within the bounding
box of the data. The cells are filled with interpolated values using the inverse weighting distance metric
with power \code{idp}. \code{nmax} determines the maximum number of neighbours when using the distance weighting.
Interpolation uses the inverse distance weight function \code{gstat} in the \code{gstat} package.
Refer to the package \code{gstat} for more details and formulae.
}
\examples{
df <- data.frame(x = runif(200),y = runif(200),z=rnorm(200))
df.gridded <- regrid(df, n1=10)
}
\keyword{distance}
\keyword{inverse}
\keyword{regrid,interpolate,}
\keyword{weighting}

