% Generated by roxygen2 (4.0.2): do not edit by hand
\name{test_image}
\alias{test_image}
\title{Create a test image}
\usage{
test_image(h = 1, r = 10, n1 = 64, n2 = 64)
}
\arguments{
\item{h}{the amplitude of the filled circle}

\item{r}{the radius of the circle (in pixesl)}

\item{n1}{image height in pixels}

\item{n2}{image width in pixels}
}
\value{
List with two elements
\describe{
\item{\code{z}}{the test image}
\item{\code{signal.grid}}{the x-y grid in long table format}
}
}
\description{
This function generates an image for test purposes. The image is that of a filled circle
at the centre.
}
\examples{
Z <- test_image()$z
}
\references{
Shen, Xiaotong, Hsin-Cheng Huang, and Noel Cressie. "Nonparametric hypothesis testing for a spatial signal." Journal of the American Statistical Association 97.460 (2002): 1122-1140.
}
\keyword{EFDR}
\keyword{image,}
\keyword{test}

